数据库换成postgersql

+++ src/modules/system/role/role.entity.ts
@@ -20,7 +20,7 @@ export class RoleEntity extends CompleteEntity {
   @ApiProperty({ description: '角色描述' })
   remark: string
 
-  @Column({ type: 'tinyint', nullable: true, default: 1 })
+  @Column({ type: 'smallint', nullable: true, default: 1 })
   @ApiProperty({ description: '状态：1启用，0禁用' })
   status: number
